+++ ViennaNET.Messaging.RabbitMQQueue/RabbitMQQueueMessageAdapter.cs
@@ -345,7 +345,7 @@ namespace ViennaNET.Messaging.RabbitMQQueue
       _logger.LogDebug(
         $"Message has been {(isSend ? "sent to" : "received from")} queue with ID:{{queueId}}{Environment.NewLine}{{message}}",
         _configuration.Id,
-        message);
+        message?.LogBody());
     }
   }
 }