+++ ViennaNET.Orm.Tests/Unit/Repositories/CommandExecutorTests.cs
@@ -1,8 +1,10 @@
-﻿using ViennaNET.Orm.Repositories;
+﻿using System.Threading;
+using System.Threading.Tasks;
 using Moq;
 using NHibernate;
 using NHibernate.Type;
 using NUnit.Framework;
+using ViennaNET.Orm.Repositories;
 using ViennaNET.Orm.Tests.Unit.DSL;
 
 namespace ViennaNET.Orm.Tests.Unit.Repositories
@@ -22,6 +24,20 @@ namespace ViennaNET.Orm.Tests.Unit.Repositories
 
       sqlQuery.Verify(x => x.SetParameter("param", 12L, TypeFactory.Basic(typeof(long).FullName)));
       sqlQuery.Verify(x => x.ExecuteUpdate(), Times.Once);
+    }
+
+    [Test]
+    public async Task ExecuteAsync_CommandWithParameters_QueryMethodCalled()
+    {
+      var sqlQuery = new Mock<ISQLQuery>();
+      var session = new Mock<ISession>();
+      session.Setup(x => x.CreateSQLQuery(It.IsAny<string>())).Returns(sqlQuery.Object);
+      var commandExecutor = new CommandExecutor<TestCommand>(session.Object);
+
+      await commandExecutor.ExecuteAsync(new TestCommand());
+
+      sqlQuery.Verify(x => x.SetParameter("param", 12L, TypeFactory.Basic(typeof(long).FullName)));
+      sqlQuery.Verify(x => x.ExecuteUpdateAsync(It.IsAny<CancellationToken>()), Times.Once);
     }
   }
 }