+++ ViennaNET.Redis.Tests/Unit/RedisPackageTests.cs
@@ -1,7 +1,10 @@
 ﻿using ViennaNET.Redis.DefaultConfiguration;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging.Abstractions;
 using NUnit.Framework;
 using SimpleInjector;
+using ViennaNET.Redis.Implementation.Default;
 
 namespace ViennaNET.Redis.Tests.Unit
 {
@@ -24,6 +27,8 @@ namespace ViennaNET.Redis.Tests.Unit
         var configuration = new ConfigurationBuilder().AddJsonFile("TestConfiguration/redis.json", optional: true)
                                                       .Build();
         _container.RegisterInstance<IConfiguration>(configuration);
+        _container.RegisterInstance<ILoggerFactory>(new NullLoggerFactory());
+        _container.Register(typeof(ILogger<>), typeof(NullLogger<>), Lifestyle.Singleton);
         _container.RegisterPackages(new[] { typeof(RedisPackage).Assembly });
         _container.Verify();